Implement defrag support for `IValue`. (#1)
The PR adds support for defrag on `IValue`. The defrag uses the provided `DefragAllocator` to reallocate all the memory of the `IValue`.

### Shared String Support

In general, it is considered easy to reallocate all the pointer of the `IValue` the only exception is the shared strings which we can not simply reallocate and free the old pointer because there might be more references that points to it.

To solve this issue we provide a way to clean the shared strings dictionary. This will cause new strings to be allocated from scratch and old shared strings will be freed when there will have no more references.

Notice that this means that now a shared string might not actually be shared between all jsons. Two json might have the same strings but with a different instance. This means that we can no longer count on pointer value for strings hash function, we must perform the hash function on the string value itself.